Cleanup react code
